[DART] handeled imports for strategy exercise 1
