+++ dart_lang/lib/b_behavioral/b_strategy/c_exercise/exercise_1/notification_strategy/notification_strategy.dart
@@ -1,0 +1,2 @@
+export 'notification_strategy_inteface.dart';
+export 'strategies/strategies.dart';
