+++ dart_lang/lib/b_behavioral/b_strategy/c_exercise/exercise_1/notification_strategy/notification_strategy_inteface.dart
@@ -1,4 +1,4 @@
-import 'package:dart_lang/b_behavioral/b_strategy/c_exercise/user.dart';
+import 'package:dart_lang/b_behavioral/b_strategy/c_exercise/exercise_1/user.dart';
 
 abstract class NotificationStrategy {
   void sendNotification(User user, String message);
