+++ dart_lang/lib/b_behavioral/b_strategy/c_exercise/exercise_1/notification_strategy/strategies/sms_notification_strategy.dart
@@ -1,0 +1,9 @@
+import 'package:dart_lang/b_behavioral/b_strategy/c_exercise/exercise_1/notification_strategy/notification_strategy_inteface.dart';
+import 'package:dart_lang/b_behavioral/b_strategy/c_exercise/exercise_1/user.dart';
+
+class SmsNotificationStrategy implements NotificationStrategy {
+  @override
+  void sendNotification(User user, String message) {
+    print("sending SMS... ");
+  }
+}
