--- dart_lang/lib/b_behavioral/b_strategy/c_exercise/notification_strategy/strategies/email_notification_strategy.dart
@@ -1,9 +1,0 @@
-import 'package:dart_lang/b_behavioral/b_strategy/c_exercise/notification_strategy/notification_strategy_inteface.dart';
-import 'package:dart_lang/b_behavioral/b_strategy/c_exercise/user.dart';
-
-class EmailNotificationStrategy implements NotificationStrategy {
-  @override
-  void sendNotification(User user, String message) {
-    print("sending Email... ");
-  }
-}
